feat(type-info): Add support for JsonTypeInfoAs.EXTERNAL_PROPERTY
This mimics the `EXTERNAL_PROPERTY` mode from Jackson to support so-called “natural” type-info properties.

If used on a propery/method/parameter level type it will look for the named type-info property at the current JSON object level (i.e. as a sibling property to the property being deserialized).

If used on a top level type it works identically to `JsonTypeInfoAs.PROPERTY`. This appears to be so that it saves repeating the type-info annotations on every usage (as long as the same “property” name is always used).