fix(minify): support minified client packages
`getArgumentNames` tests for specific code pieces but assumes whitespace that may be removed by minification.  The tests are changed to use regex’s that make the space optional.

After minification code arises like the following:
```
class{constructor(e,t,r,n,i,a,o){this.id=e,this.aliases=t,this.caCertificate=r,this.hash=n,this.inception=i,this.deviceCount=a,this.data=o}}
```
Notice the lack of `extends` after `class`. This causes the `class extends` test to fail. This regex was altered to make the `extends` portion optional which causes replacement with `class JacksonClass`; as is required.